+++ src/App.js
@@ -9,7 +9,7 @@ import Admin from "./components/Admin"
 import Editor from "./components/Editor"
 import Missing from "./components/Missing"
 import Lounge from "./components/Lounge"
-import RequireAuth from "./components/RequireLogin"
+import RequireAuth from "./components/RequireAuth"
 
 function App() {
   return (
@@ -20,10 +20,19 @@ function App() {
         <Route path="linkpage" element={<LinkPage />} />
         <Route path="unauthorized" element={<Unauthorized/>} />
         
-        <Route element={<RequireAuth />}>
+        <Route element={<RequireAuth allowedRoles={[2001]} />}>
           <Route path="/" element={<Home />} />
+        </Route>
+
+        <Route element={<RequireAuth allowedRoles={[1984]} />}>
           <Route path="editor" element={<Editor />} />
+        </Route>
+
+        <Route element={<RequireAuth allowedRoles={[5150]} />}>
           <Route path="admin" element={<Admin />} />
+        </Route>
+
+        <Route element={<RequireAuth allowedRoles={[1984, 5150]} />}>
           <Route path="lounge" element={<Lounge />} />
         </Route>
